+++ lib/rubocop.rb
@@ -207,6 +207,7 @@ require 'rubocop/cop/rails/default_scope'
 require 'rubocop/cop/rails/has_and_belongs_to_many'
 require 'rubocop/cop/rails/output'
 require 'rubocop/cop/rails/read_attribute'
+require 'rubocop/cop/rails/scope_args'
 require 'rubocop/cop/rails/validation'
 
 require 'rubocop/formatter/base_formatter'